Moved DB stuff into a separate package in preparation for the DB work
