+++ src/main/scala/thurloe/service/ThurloeServiceActor.scala
@@ -3,6 +3,7 @@ package thurloe.service
 import akka.actor.{ActorRefFactory, ActorLogging}
 
 import spray.routing._
+import thurloe.database.{ThurloeDatabaseConnector, DataAccess}
 import scala.util.{Try, Success}
 
 // we don't implement our route structure directly in the service actor because
@@ -33,13 +34,3 @@ case class HostedResource(resource: String, resourceDirectory: Option[String], p
   }
 }
 
-case object ThurloeDatabaseConnector extends DataAccess {
-  
-  // TODO: Fill THESE in with database access implementation.
-  def keyLookup(key: String) = Success(KeyValuePair("yek", "eulav"))
-  def collectAll() = Success(Seq(
-    KeyValuePair("key1", "Bob Loblaw's Law Blog"),
-    KeyValuePair("key2", "Blah blah blah blah blah")))
-  def setKeyValuePair(keyValuePair: KeyValuePair): Try[Unit] = Success(())
-  def deleteKeyValuePair(key: String): Try[Unit] = Success()
-}
